Added Planet class and methods.
Uses a list called planetList to store a list of Planet objects.
First, asks the user to enter the number of planets that should be generated.
Second, uses a for loop to generate the planet objects as well as their asscociative attributes.
Third, enters a while loop after asking the user to enter the specific number of a planet
or press 'a' to see the stats of all planets. Entering anything else will close the program.

The Planet object has the following methods:

PrintPlanetName() - Prints the planet's name.

PrintPlanetSize() - Prints the planet's size.

PrintPlanetPop() - Prints the planet's population.

PrintPlanetType() - Prints the planet's main biome type in the appropriate color.

GrowPop(bool message) - Increases the planet's population by 15%. If the message parameter is
set to true, a message appears that tells the user about the change along with the old and new values.