Reformatted program and added functionality
I rewrote the entire program logic that determined the program flow.
It now uses a 'do-while' loop and a switch statement that determines the
route the user will be able to take through the program. Increases
usability and readability.

I added an additional six possible first and last name segments as well
as planet types the program can choose from. I also introduced some
code that will eventually allow for generated planets to be gas giants,
although currently it is unimplemented.

I added a new class called Methods.cs that holds most of the common
methods used in the main program.