Added Unit Converter
Added an advanced unit converter as well as a class to hold a majority of the functions relating to it. It allows a user to convert miles, kilometers, light years, or astronomical units to any other unit. To aid in its relevance, I also added DISTANCE TO STAR as a property for the randomly generated planets.